+++ CHANGELOG.md
@@ -2,12 +2,19 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
-## [Unreleased]
+## [0.0.2] - 2017-08-19
+### Added
+- Create git repo and first commit for fresh project
+- Django authentication supported (only login and logout now)
+
 ### Fixed
 - Modification of MIDDLAWARE settings in settings_local.py fixed
 
-### Added
-- Create git repo and first commit for fresh project
+## Changed
+- Use settings_local.py config file first if exists (not need set
+  DJANGO_SETTINGS_MODULE)
+- Documentation minor update
+
 
 
 ## [0.0.1] - 2017-08-16